services/worker/requirements.txt
celery==5.3.4
redis==5.0.1
pika==1.3.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
structlog==24.1.0
flower==2.0.1
